+++ horovod/tensorflow/__init__.py
@@ -43,6 +43,7 @@ from horovod.tensorflow.util import _executing_eagerly, _make_subgraph, _cache, 
 from horovod.tensorflow.mpi_ops import join
 from horovod.tensorflow.sync_batch_norm import SyncBatchNormalization
 from horovod.tensorflow.gradient_aggregation import LocalGradientAggregationHelper
+from horovod.common.util import support_non_legacy_keras_optimizers
 
 import tensorflow as tf
 _IS_TF2 = version.parse(tf.__version__) >= version.parse('2.0.0')
@@ -680,13 +681,14 @@ if _LegacyOptimizer is not None:
                     rank=rank(),
                     optimizer_type=LocalGradientAggregationHelper._OPTIMIZER_TYPE_LEGACY,
                     process_set=process_set,
-                    scale_local_gradients=scale_local_gradients
+                    scale_local_gradients=scale_local_gradients,
+                    name=name,
                 )
 
         def register_local_var(self, var):
             """Registers a source/variable as worker local. Horovod will not perform any global
             operations on gradients corresponding to these sources and will instead return the local
-            gradient."""    
+            gradient."""
             if self._agg_helper:
                 self._agg_helper.register_local_var(var)
             elif _IS_TF2:
@@ -995,7 +997,9 @@ def DistributedOptimizer(optimizer, name=None, use_locking=False, device_dense='
             process_set=process_set,
             scale_local_gradients=scale_local_gradients
         )
-    elif isinstance(optimizer, tf.keras.optimizers.Optimizer):
+    elif (isinstance(optimizer, tf.keras.optimizers.Optimizer) or
+          (not support_non_legacy_keras_optimizers(tf.keras) and
+           isinstance(optimizer, tf.keras.optimizers.legacy.Optimizer))):
         if op == Adasum:
             raise ValueError('op == Adasum is not supported yet with Keras')